third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/events/event-listener-moving-documents-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Moving a Mutation Event listener between documents should keep it active

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS addedCalls is 1
PASS addedCalls is 2
PASS removedCalls is 1
PASS successfullyParsed is true

TEST COMPLETE
